+++ src/main/java/br/com/projeto/sistemadeavaliacao/model/Pergunta.java
@@ -13,6 +13,6 @@ public class Pergunta {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String avaliacao;
+    private String textoPergunta;
 
 }
