+++ src/main/java/br/com/projeto/sistemadeavaliacao/model/Secretaria.java
@@ -2,21 +2,9 @@ package br.com.projeto.sistemadeavaliacao.model;
 
 import javax.persistence.Entity;
 import javax.persistence.PrimaryKeyJoinColumn;
-import javax.validation.constraints.NotEmpty;
 
 @Entity
 @PrimaryKeyJoinColumn(name = "userID")
 public class Secretaria extends Usuario {
-
-	@NotEmpty
-	private String nome;
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
 
 }
